+++ src/main/java/io/legado/app/data/entities/BookGroup.kt
@@ -7,7 +7,7 @@ package io.legado.app.data.entities
 //@Entity(tableName = "book_groups")
 data class BookGroup(
 //        @PrimaryKey
-        var groupId: Int = 0,
+        var groupId: Long = 0,
         var groupName: String = "",
         var cover: String? = null,
         var order: Int = 0,
